+++ tools/fastdump/sorter.h
@@ -50,7 +50,7 @@ rc_t execute_lookup_production( KDirectory * dir,
                                 size_t cursor_cache,
                                 size_t buf_size,
                                 size_t mem_limit,
-                                size_t num_threads,
+                                uint32_t num_threads,
                                 bool show_progress );
 
 #ifdef __cplusplus
